+++ lib/rspec/mocks/message_expectation.rb
@@ -452,10 +452,8 @@ module RSpec
         super(error_generator, expectation_ordering, expected_from, method_double, 0, {}, &implementation)
       end
 
-      # no-op
-      # @deprecated and_return is not supported with negative message expectations.
       def and_return(*)
-        RSpec.deprecate "and_return with should_not_receive"
+        raise "and_return is not supported with negative message expectations"
       end
 
       # @private